+++ CHANGELOG.rst
@@ -32,6 +32,8 @@ spyne-2.10.0
    You can set ``ctx.out_document`` to have the return value from user funtion
    ignored. You can also set ``ctx.out_string`` to have the ``ctx.out_document``
    ignored as well.
+ * Added as_time_zone support to DateTime. It calls
+   ``.astimezone(as_time_zone).replace(tzinfo=None)`` on native values.
 
 spyne-2.9.3
 -----------